Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -2,7 +2,7 @@
 int main () {
     int a;
     scanf("%d",&a);
-    float b = r*r*3.14 ;
+    int b = r*r*3.14 ;
     printf("Area: %d",b);
     return 0;
 }
